+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed toggle volumetric lighting in the light UI
 - Fixed post-processing history reset handling rt-scale incorrectly
 - Fixed crash with terrain and XR multi-pass
+- Fixed ShaderGraph material synchronization issues
 
 ### Changed
 - Update Wizard layout.